# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SYNCVERSE AI - Real-time campus connection engine
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}/squad-suggestions:
    get:
      operationId: getSquadSuggestionsForUser
      tags: [squads]
      summary: AI-generated squad suggestions for a user
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Suggested squads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Squad"
  /squads/{squadId}/join:
    post:
      operationId: joinSquad
      tags: [squads]
      summary: Join a squad
      parameters:
        - in: path
          name: squadId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinSquadInput"
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Squad"
...
components:
  schemas:
...
    CommunityZone:
      type: string
      enum: [career, startup, study, social, creative, fitness, research]
    Timeframe:
      type: string
      enum: [now, soon, later]
    EnergyLevel:
      type: string
      enum: [browsing, exploring, building]
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        college:
          type: string
        major:
          type: string
        intent:
          type: string
        timeframe:
          $ref: "#/components/schemas/Timeframe"
        energyLevel:
          $ref: "#/components/schemas/EnergyLevel"
        zone:
          $ref: "#/components/schemas/CommunityZone"
        avatarColor:
          type: string
        createdAt:
          type: string
      required: [id, name, college, major, intent, timeframe, energyLevel, zone, avatarColor, createdAt]
...
    Squad:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        purpose:
          type: string
        firstAction:
          type: string
        suggestedMeetup:
          type: string
        zone:
          $ref: "#/components/schemas/CommunityZone"
        members:
          type: array
          items:
            $ref: "#/components/schemas/User"
        createdAt:
          type: string
      required: [id, name, purpose, firstAction, suggestedMeetup, zone, members, createdAt]
    JoinSquadInput:
      type: object
      properties:
        userId:
          type: string
      required: [userId]
...